Make `write_matrix` separately compiled
