+++ include/binsparse/convert_matrix.h
@@ -9,8 +9,8 @@
 #include <assert.h>
 #include <binsparse/matrix.h>
 
-bsp_matrix_t bsp_convert_matrix(bsp_matrix_t matrix,
-                                bsp_matrix_format_t format) {
+static inline bsp_matrix_t bsp_convert_matrix(bsp_matrix_t matrix,
+                                              bsp_matrix_format_t format) {
   // Throw an error if matrix already in desired format.
   if (matrix.format == format) {
     assert(false);
