+++ src/CMakeLists.txt
@@ -4,4 +4,5 @@
 
 target_sources(binsparse-rc PRIVATE
   src/read_matrix.c
+  src/write_matrix.c
 )
